% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMergely.R
\name{shinyMergely}
\alias{shinyMergely}
\title{Shiny Mergely}
\usage{
shinyMergely()
}
\description{
Launch a Shiny app allowing to compare and merge two files.
}
\examples{
if(interactive()){
  shinyMergely()
}
}
